requirements.txt
accelerate==0.25.0
anyio==4.1.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boilerpy3==1.0.7
boto3==1.34.47
botocore==1.34.47
bs4==0.0.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.3.1
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.0
executing==2.0.1
farm-haystack==1.22.1
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.17.3
idna==3.6
inflect==7.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy-imports==0.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.13
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.41
prompthub-py==4.0.0
protobuf==4.25.1
psutil==5.9.6
psycopg==3.1.17
psycopg-binary==3.1.17
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.13
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.13.2
rubicon-objc==0.4.7
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.40.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.14.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
transformers==4.34.1
typing_extensions==4.9.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==3.0.1
whitenoise==6.4.0
accelerate==0.25.0
anyio==4.1.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boilerpy3==1.0.7
boto3==1.34.47
botocore==1.34.47
bs4==0.0.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
decorator==5.1.1
distro==1.9.0
Django==4.2.4
django-cors-headers==4.3.1
django-environ==0.10.0
djangorestframework==3.14.0
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.0
executing==2.0.1
farm-haystack==1.22.1
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.17.3
idna==3.6
inflect==7.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy-imports==0.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
num2words==0.5.13
numpy==1.26.2
openai==1.17.0
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.41
prompthub-py==4.0.0
protobuf==4.25.1
psutil==5.9.6
psycopg==3.1.17
psycopg-binary==3.1.17
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.13
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.13.2
rubicon-objc==0.4.7
s3transfer==0.10.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.19.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.40.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.14.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
transformers==4.34.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==3.0.1
whitenoise==6.4.0
wsproto==1.2.0
